+++ src/pages/BridgePage/BridgeForm.tsx
@@ -25,7 +25,9 @@ const BridgeForm = () => {
     destinationChain,
     destinationChainOptions,
     setDestinationChain,
-    switchOriginAndDestination
+    switchOriginAndDestination,
+    originChainIsEvm,
+    destinationChainIsEvm
   } = useBridgeData();
 
   useEffect(() => {
@@ -40,8 +42,6 @@ const BridgeForm = () => {
     }
   };
 
-  const originChainIsEvm = originChain?.xcmAdapter.chain.type === 'ethereum';
-  const destinationChainIsEvm = destinationChain?.xcmAdapter.chain.type === 'ethereum';
   const shouldShowDestinationInput = originChainIsEvm || destinationChainIsEvm;
 
   return (