+++ src/pages/SendPage/SendContext/index.tsx
@@ -17,7 +17,7 @@ import SEND_ACTIONS from './sendActions';
 import sendReducer, { buildInitState } from './sendReducer';
 import {
   updateHistoryEventStatus,
-  removePendingHistoryEvent,
+  removePendingHistoryEvent
 } from 'utils/persistence/privateTransactionHistory';
 import { HISTORY_EVENT_STATUS } from 'types/HistoryEvent';
 
@@ -324,18 +324,17 @@ export const SendContextProvider = (props) => {
   };
 
   // Returns true if the current tx would cause the user to go below a
-  // recommended min fee balance of 1. This helps prevent users from
+  // recommended min fee balance of 150. This helps prevent users from
   // accidentally becoming unable to transact because they cannot pay fees
   const txWouldDepleteSuggestedMinFeeBalance = () => {
     if (
       senderAssetCurrentBalance?.assetType.isNativeToken &&
-      !senderAssetCurrentBalance?.assetType.isPrivate &&
       senderAssetTargetBalance?.assetType.isNativeToken &&
-      !senderAssetTargetBalance?.assetType.isPrivate
+      isToPrivate()
     ) {
       const SUGGESTED_MIN_FEE_BALANCE = Balance.fromBaseUnits(
         AssetType.Native(config),
-        1
+        150
       );
       const balanceAfterTx = senderAssetCurrentBalance.sub(
         senderAssetTargetBalance